Improve the responsiveness of the contents and lists pages of the help center
