+++ helpcenter/src/app/list/page.tsx
@@ -1,11 +1,11 @@
-import Content from '@/components/components/pages/Content'
+import List from '@/components/components/pages/List'
 import NavBar from '@/components/components/pages/NavBar'
 
 export default function page() {
   return (
     <div>
         < NavBar />
-        < Content />
+        < List />
     </div>
   )
 }
